backend/requirements.txt
argon2-cffi
dbutils
fastapi
Jinja2
python-jose
langchain
langchain_community
langchain_core
langserve
pydantic==1.*
pyodbc
python-dotenv
PyYAML
SQLAlchemy
sqlglot
uvicorn
python-multipart
sse_starlette

sentence-transformers
chromadb
mysql_connector_repackaged
psycopg2-binary
